+++ src/driver/oracle/OracleDriver.ts
@@ -450,7 +450,7 @@ export class OracleDriver implements Driver {
     /**
      * Creates generated map of values generated or returned by database after INSERT query.
      */
-    createGeneratedMap(metadata: EntityMetadata, insertionResult: any) {
+    createGeneratedMap(metadata: EntityMetadata, insertValue: ObjectLiteral, insertResult: any) {
         return undefined;
     }