+++ src/driver/sqlserver/SqlServerDriver.ts
@@ -469,7 +469,7 @@ export class SqlServerDriver implements Driver {
     /**
      * Creates generated map of values generated or returned by database after INSERT query.
      */
-    createGeneratedMap(metadata: EntityMetadata, insertResult: any) {
+    createGeneratedMap(metadata: EntityMetadata, insertValue: ObjectLiteral, insertResult: any) {
         if (!insertResult || !insertResult[0])
             return undefined;