+++ src/metadata/EntityMetadata.ts
@@ -513,20 +513,21 @@ export class EntityMetadata {
     }
 
     /**
-     * Finds column with a given property path.
+     * Finds columns with a given property path.
+     * Relations can contain multiple columns.
      */
-    findColumnWithPropertyPath(propertyPath: string): ColumnMetadata|undefined {
+    findColumnsWithPropertyPath(propertyPath: string): ColumnMetadata[] {
         const column = this.columns.find(column => column.propertyPath === propertyPath);
         if (column)
-            return column;
+            return [column];
 
         // in the case if column with property path was not found, try to find a relation with such property path
         // if we find relation and it has a single join column then its the column user was seeking
         const relation = this.relations.find(relation => relation.propertyPath === propertyPath);
-        if (relation && relation.joinColumns.length === 1)
-            return relation.joinColumns[0];
+        if (relation && relation.joinColumns)
+            return relation.joinColumns;
 
-        return undefined;
+        return [];
     }
 
     /**