+++ test/functional/embedded/embedded-many-to-one-case3/embedded-many-to-one-case3.ts
@@ -368,7 +368,7 @@ describe("embedded > embedded-many-to-one-case3", () => {
                 .where("post.id = :id", { id: 2 })
                 .getOne();
 
-            expect(loadedPost!.counters.likedUser).to.be.empty;
+            expect(loadedPost!.counters.likedUser).to.be.null;
         })));
 
     });