+++ test/functional/embedded/embedded-many-to-one-case4/embedded-many-to-one-case4.ts
@@ -377,7 +377,7 @@ describe("embedded > embedded-many-to-one-case4", () => {
                 .where("post.id = :id", { id: 2 })
                 .getOne();
 
-            expect(loadedPost!.counters.likedUser).to.be.empty;
+            expect(loadedPost!.counters.likedUser).to.be.null;
         })));
 
     });