+++ test/functional/embedded/embedded-many-to-one-case5/embedded-many-to-one-case5.ts
@@ -382,7 +382,7 @@ describe("embedded > embedded-many-to-one-case5", () => {
                 .where("post.id = :id", { id: 2 })
                 .getOne();
 
-            expect(loadedPost!.counters.likedUser).to.be.empty;
+            expect(loadedPost!.counters.likedUser).to.be.null;
         })));
 
     });